feat: add conditional rendering in Header Component to decide which menu will display
In Header Component it's necessary to render properly which menu should display to the user. In that
way, if the viewport width is more than 750px, MenUDesktop should be rendered, if not, the
MenuMobile will be displayed. Therefore, an UseState was created to store the viewport width and an
UseEffect to update the useState. After it, a conditional rendering was add to decide which menu
will be displayed.